Extrace treewalk interpriter
