+++ Cargo.toml
@@ -18,6 +18,7 @@ parser = { path = "./crates/parser" }
 regex = "1.4.6"
 serde = { version = "1.0.125", features = ["derive"] }
 slotmap = { version = "1.0.3", features = ["serde"] }
+treewalk = { path = "./crates/treewalk" }
 value = { path = "./crates/value" }
 wasm-bindgen = "0.2.74"
 
@@ -37,4 +38,5 @@ members = [
     "./crates/parser",
     "./crates/diagnostics",
     "./crates/value",
+    "./crates/treewalk"
 ]
