+++ crates/treewalk/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "treewalk"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+codespan-reporting = "0.11.1"
+diagnostics = { path = "../diagnostics" }
+eyre = "0.6.5"
+parser = { path = "../parser" }
+value = { path = "../value" }
